moved most of the logic to src/lib.rs
